(#237)refactor: 신고 service interface 구현했습니다.
